this uploads everything including the pics now. structure is as following:
css/<css-foldername>/<cssfile>
css/<css-foldername>/images/<image.something>

with the exception of the sub styles which end up here:

css/sub/<css-foldername>/<cssfile>
sub/<css-foldername>/images/<image.something>

killing the sub could result in some nasty clashing of css files hence it's better to leave it.

a real problem is that currently the script throws when the folder is not existing. this needs some real fixing.

git-svn-id: svn://dev.anidb.net/svn/trunk@19 bed53a55-4da3-433c-baf4-d4f70f2fd34a