Fixing several inconsistencies in the ajax anime page.
[1276] 'add to mylist' button on notidy generated pages doesn't use quick add
[1275] ed2k links in notifies not respecting custom filenames

Anyways added, i expect, the right hooks for the missing stuff that could call the anime page, most notably eid and gid expansions.
And YES, i know that it redraws the tables again, it's either that or nothing, it's too much of a pain to do updates row for row, while having to deal with missing rows and everything.

git-svn-id: svn://dev.anidb.net/svn/trunk@1807 bed53a55-4da3-433c-baf4-d4f70f2fd34a