wasn't quite working as i expected it...
Anyways, anime info popup for mylist now has a icon for explicit mouse over. As in, if you are overing the icon you want to see the description and so on.

This setting is activated in your profile, where you can also set the default size of the thumbnail displayed.

git-svn-id: svn://dev.anidb.net/svn/trunk@1826 bed53a55-4da3-433c-baf4-d4f70f2fd34a