Did a runtime check for ffmpeg libs that works, if the app can't load them avparsing will be disabled.
also got a conversion table of internal codecs to anidb format to take care of the memory issues (still) now i'm getting no info for some stuff that should be there.

This means that i will need to find something better than jna for this.

will work on the preferences for the gui version next.

git-svn-id: svn://dev.anidb.net/svn/trunk@2009 bed53a55-4da3-433c-baf4-d4f70f2fd34a