A one-liine fix to keep WebAom from getting confused if a UDP packet is dropped (or if the server is busy and discards it).
This is not a perfect fix.  In /most/ cases, it will result in one file not being renamed.  If the timing is JUST right, the applet will remain out of sync with the server, and no subsequent files will be renamed.

However, in neither case should a file be /incorrectly/ renamed.

git-svn-id: svn://dev.anidb.net/svn/trunk@2850 bed53a55-4da3-433c-baf4-d4f70f2fd34a