Added JSON support for ajax.js.
For it to work correctly the correct method=json needs to be used.
This lib also supports a basic fall back (without JSON revivers) for legacy browsers.

A working example:

xhttpRequestFetch(xhttpRequest(), someURL, function(reply) { alert(reply); }, null, "json")

Note that the null is for the override, read the inline documentation for more help.

git-svn-id: svn://dev.anidb.net/svn/trunk@4252 bed53a55-4da3-433c-baf4-d4f70f2fd34a