new logic for the search results. if the user has set a default language for the given search type and we have a match for that language display it.
the order is something like this:

Search results:
Anime[A] > Title [Synm][pl]
Anime[A] > Title [Synm][en]
Anime[A] > Title [Official][ja]
Anime[A] > Title [Short][x-jat]
Anime[A] > Title [Main][de]
Anime[A] > Title [Synm][pt]
Anime[B] > Title [Synm][en]
Anime[B] > Title [Official][en]
Anime[B] > Title [Official][x-jat]
Anime[B] > Title [Short][x-jat]

Imagine that the user default anime language is x-jat, in that case, the results presented to the user will be:

Anime[A] > Title [Main][de]
Anime[B] > Title [Official][x-jat]

Remember, only titles that contain the search query will be returned. So the german title for anime A contains the search query. We are returning the Main title instead of the Official title that is jap transcript because
the main title contains the query and that is the current behaviour of searches in any case.
The order for picking results is:
Main > Official > Symn > Short (for anime)
Main > whatever (in the order the types were added)

git-svn-id: svn://dev.anidb.net/svn/trunk@4262 bed53a55-4da3-433c-baf4-d4f70f2fd34a