language matching now goes like this:
main title exists and contains the search query?
yes => that's our title
no?
for (type in: "Official", "synm", "short") {
do we have a default title query match on the type?
yes => that's our title
no?
for (lang in: "japanese", "japanese-transcript", "english") {
do we have a title query match for that language?
yes => that's our title
no => continue to next lang
}
did we found a title in the cycle above?
yes => nothing more to do
no => grab the first title available and that is it
}

git-svn-id: svn://dev.anidb.net/svn/trunk@4271 bed53a55-4da3-433c-baf4-d4f70f2fd34a